consolidate provider reducer enums
