+++ packages/react-paypal-js/src/v6/context/BraintreePayPalContext.test.ts
@@ -1,8 +1,8 @@
+import { braintreeReducer } from "./BraintreePayPalContext";
 import {
-    braintreeReducer,
     BRAINTREE_DISPATCH_ACTION,
-} from "./BraintreePayPalContext";
-import { INSTANCE_LOADING_STATE } from "../types/ProviderEnums";
+    INSTANCE_LOADING_STATE,
+} from "../types/ProviderEnums";
 
 import type {
     BraintreePayPalState,
